pot bucket choice fix

+++ fullnode/src/main/java/io/coti/fullnode/services/PotService.java
@@ -34,7 +34,7 @@ public class PotService extends BaseNodePotService {
 
         int trustScore = transactionData.getRoundedSenderTrustScore();
 
-        int bucketChoice = (int) (Math.ceil(trustScore / 10) * 10);
+        int bucketChoice = (int) (Math.ceil((double) trustScore / 10) * 10);
         if (queuesPot.get(bucketChoice) == null) {
             throw new IllegalArgumentException("Illegal trust score");
         }
